fix: harden bundler overrides against prototype pollution
- Use Object.prototype.hasOwnProperty.call instead of 'in' operator
- Use structuredClone to snapshot overrides at module load time
- Refactor bundler tests into separate files so __bundlerPathsOverrides
  is set before pino is loaded
