+++ lib/transport.js
@@ -71,6 +71,8 @@ function flush (stream) {
   stream.flushSync()
 }
 
+const bundlerOverrides = structuredClone(Object.prototype.hasOwnProperty.call(globalThis, '__bundlerPathsOverrides') ? globalThis.__bundlerPathsOverrides : {})
+
 function transport (fullOptions) {
   const { pipeline, targets, levels, dedupe, worker = {}, caller = getCallers(), sync = false } = fullOptions
 
@@ -80,9 +82,6 @@ function transport (fullOptions) {
 
   // Backwards compatibility
   const callers = typeof caller === 'string' ? [caller] : caller
-
-  // This will be eventually modified by bundlers
-  const bundlerOverrides = '__bundlerPathsOverrides' in globalThis ? globalThis.__bundlerPathsOverrides : {}
 
   let target = fullOptions.target
